Personalized-ChatBot/requirements.txt
streamlit==1.31.0
langchain==0.1.3
openai==1.9.0
langchain-openai==0.0.3
python-dotenv==1.0.0
PyPDF2==3.0.1
faiss-cpu==1.7.4